FIX-2120 move log worker tools to new file
