+++ Src/WitsmlExplorer.Api/Workers/LogWorkerTools.cs
@@ -1,0 +1,48 @@
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+using Microsoft.Extensions.Logging;
+
+using Witsml;
+using Witsml.Data;
+using Witsml.Extensions;
+using Witsml.ServiceReference;
+
+using WitsmlExplorer.Api.Jobs.Common.Interfaces;
+using WitsmlExplorer.Api.Query;
+using WitsmlExplorer.Api.Services;
+
+namespace WitsmlExplorer.Api.Workers
+{
+    public static class LogWorkerTools
+    {
+        public static async Task<WitsmlLog> GetLog(IWitsmlClient client, IObjectReference logReference, ReturnElements optionsInReturnElements)
+        {
+            WitsmlLogs logQuery = LogQueries.GetWitsmlLogById(logReference.WellUid, logReference.WellboreUid, logReference.Uid);
+            WitsmlLogs result = await client.GetFromStoreAsync(logQuery, new OptionsIn(optionsInReturnElements));
+            return !result.Logs.Any() ? null : result.Logs.First();
+        }
+
+        public static async Task<WitsmlLogData> GetLogDataForCurve(IWitsmlClient witsmlClient, WitsmlLog log, string mnemonic, ILogger logger)
+        {
+            await using LogDataReader logDataReader = new(witsmlClient, log, mnemonic.AsSingletonList(), logger);
+            List<WitsmlData> data = new();
+            WitsmlLogData logData = await logDataReader.GetNextBatch();
+            var mnemonicList = logData?.MnemonicList;
+            var unitList = logData?.UnitList;
+            while (logData != null)
+            {
+                data.AddRange(logData.Data);
+                logData = await logDataReader.GetNextBatch();
+            }
+
+            return new WitsmlLogData
+            {
+                MnemonicList = mnemonicList,
+                UnitList = unitList,
+                Data = data
+            };
+        }
+    }
+}
